Refactor JAX PME testing with updated Coulomb callback
- Removed the obsolete `_fake_pme` function and replaced it with `_fake_coulomb_callback` to better simulate Coulomb interactions in tests.
- Updated test cases to utilize the new callback, ensuring compatibility with the latest JAX PME enhancements.
- Improved clarity and maintainability of the test suite by aligning with recent changes in long-range interaction handling.

+++ mmml/interfaces/pycharmmInterface/mm_energy_forces.py
@@ -556,7 +556,7 @@ def _wrap_mm_fn_with_jax_pme(
         if box_override is not None:
             return _box_length_from_cell_jax(box_override)
         if static_box_L is not None:
-            return jnp.asarray(static_box_L, dtype=jnp.float32)
+            return jnp.asarray(static_box_L)
         raise ValueError("jax_pme requires pbc_cell or box_override")
 
     def _add_pme(
